requirements.txt
Django==3.2.14
djangorestframework==3.13.1
pyyaml==6.0
uritemplate==4.1.1
factory_boy==3.2.1
Faker==13.15.1
